+++ src/pat/navigation/index.html
@@ -4,6 +4,7 @@
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <title>Navigation demo</title>
     <link rel="stylesheet" href="/style/common.css" type="text/css" />
+    <script src="/bundle-vendors.js" type="text/javascript" charset="utf-8"></script>
     <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
     <style type="text/css" media="screen">
 
@@ -155,6 +156,5 @@
         Content for tab 1.2
       </div>
     </section>
-
   </body>
 </html>
